requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cryptography==3.4.8
distro==1.9.0
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.13.0
django-phonenumber-field==8.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
nltk==3.9.4
numpy==2.4.4
pandas==3.0.2
phonenumbers==9.0.26
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
regex==2026.3.32
requests==2.33.1
scikit-learn==1.7.1
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
textblob==0.20.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
